# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto
  description: >
    This api allows user to create their account and get connected with their
    friends by following them to see their photos and react on them.
  version: '1'
...
paths:
...
  /user/{username}/photos/likes:
    parameters:
      - name: username 
        in: path 
        description: feed of the user
        required: true
        schema:
          $ref: '#/components/schemas/username'
    post:
      security:
        - bearerAuth: [ ] 
      tags: 
        - Photo Operation
      summary: liking photo
      description: sending a like to own's photo or someone's photo whom the user follows 
      operationId: likePhoto
      requestBody:
        description: sending like to a photo
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/like'
        required: true
      responses:
        '201':
          description: successfully liked the photo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/photos'
        '204':
          description: already liked photo 
    
    delete:
      security:
        - bearerAuth: [ ] 
      tags:
        - Photo Operation
      summary: removing like 
      description: removing like from a photo which is already liked 
      operationId: unlikePhoto
      responses:
        '204':
          description: successfully disliked the photo
  /user/{username}/photos/comment:
    parameters:
        - name: username 
          in: path 
          description: feed of the user
          required: true
          schema:
            $ref: '#/components/schemas/username'
    post:
      security:
        - bearerAuth: [ ] 
      tags:
        - Photo Operation
      summary: commenting on a photo
      description: commenting on a photo which is either uploaded by the user himself or by someone whom user follows
      operationId: commentPhoto
      requestBody:
        description: commenting on a photo from user's feed or user's own photo
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/comments'
        required: true
      responses:
        '201':
          description: successfully commented on the photo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/photos'
    delete:
      security:
        - bearerAuth: [ ] 
      tags:
        - Photo Operation
      summary: removing comment
      description: removing comment from a photo which is already commented which means only the author can delete his comment 
      operationId: uncommentPhoto
      responses:
        '204':
          description: successfully uncommented the photo
...
components:
  securitySchemes:
    bearerAuth:         
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    comments:
      description: comments 
      type: object
      properties:
        photo_username_number:
          description: photoid of the photo
          type: string
          pattern:  '^.*?$'
          minLength: 0
          maxLength: 50
          example: maria1205
        commenter_username:
          description: username of the commenter 
          type: string
          pattern:   '^(?!.*_).*$'
          minLength: 0
          maxLength: 30
          example: abcdef012345
        comment:
          description: comment 
          type: string
          pattern:  '^.*?$'
          minLength: 5
          maxLength: 50
          example: beautiful
    like:
      type: object
      description: like on the photo
      properties:
        photo_username_number:
          description: photoid of the photo
          type: string
          pattern:  '^.*?$'
          minLength: 0
          maxLength: 50
          example: maria1205
        liked_by:
          description: id of the liker
          type: string
          pattern:  '^(?!.*_).*$'
          minLength: 0
          maxLength: 30
          example: abcdef012345
        like:
          description: like or unlike 
          type: boolean
    username:
      type: object
      description: username 
      properties:
        identifier:
          description: username
          type: string
          minLength: 0
          maxLength: 30
          example: abcdef012345
          pattern: '^(?!.*_).*$'
    photos:
      description: 'it gives photos of particular user with photo number and all details about the photo'
      type: array
      minItems: 5
      maxItems: 7
      items:
        oneOf:
        - type: string
          description: image in form of binary bits 
          format: binary
        - 
          type: string 
          pattern:   '^.*?$'
          description: userid of the uploader concataned with number of the photo of that user to make it a unique id for photo
          minLength: 0
          maxLength: 40
          example: 'maria_2'
        - type: string 
          description:  date and time of photo
          format: date-time
          example: 2017-07-21T17:32:28Z
        - type: integer 
          description: number of likes
          minimum: 0
          example: 50
        - type: integer
          description: number of comments
          minimum: 0
          example: 23